Create total attention results component to display chart and handle logic
