+++ mcweb/frontend/src/features/search/Search.jsx
@@ -12,6 +12,7 @@ import SimpleSearch from './query/SimpleSearch';
 import SampleStories from './results/SampleStories';
 import { setSearchTime, removeSelectedMedia } from './query/querySlice';
 import TotalAttentionChart from './results/TotalAttentionChart';
+import TotalAttentionResults from './results/TotalAttentionResults';
 import CountOverTimeResults from './results/CountOverTimeResults';
 import MediaPicker from './query/media-picker/MediaPicker';
 import urlSerializer from './util/urlSerializer';
@@ -136,7 +137,7 @@ export default function Search() {
       <div className="search-results-wrapper">
         <div className="container">
           <CountOverTimeResults />
-          <TotalAttentionChart />
+          <TotalAttentionResults />
           <SampleStories />
         </div>
       </div>
